Consolidate styling of featured carousels
- These should use the same styles and we don't want a background on one
  but not the other
