+++ src/components/carousel/FeaturedPlaylists.test.tsx
@@ -2,12 +2,12 @@ import { FakeBoclipsClient } from 'boclips-api-client/dist/test-support';
 import { CollectionFactory } from 'src/testSupport/CollectionFactory';
 import { render } from '@testing-library/react';
 import React from 'react';
-import FeaturedPlaylists from 'src/components/featuredPlaylists/FeaturedPlaylists';
 import { BoclipsClientProvider } from 'src/components/common/providers/BoclipsClientProvider';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { VideoFactory } from 'boclips-api-client/dist/test-support/VideosFactory';
 import { PromotedForProduct } from 'boclips-api-client/dist/sub-clients/collections/model/PromotedForProduct';
+import FeaturedPlaylists from 'src/components/carousel/FeaturedPlaylists';
 
 describe(`PromotedPlaylists`, () => {
   it(`renders all playlists promoted for product`, async () => {
