+++ src/components/carousel/PlaylistSlide.tsx
