+++ src/views/home/HomeView.tsx
@@ -7,7 +7,6 @@ import HomeSearchHero from 'src/resources/icons/home-illustration.svg';
 import HomeSearchHeroClassroom from 'src/resources/icons/classroom/home-search-hero.svg';
 import { useNavigate } from 'react-router-dom';
 import NewNavbar from 'src/components/layout/Navbar';
-import FeaturedPlaylists from 'src/components/featuredPlaylists/FeaturedPlaylists';
 import FeaturedVideos from 'src/components/carousel/FeaturedVideos';
 import { useMediaBreakPoint } from '@boclips-ui/use-media-breakpoints';
 import { useEmailVerified } from 'src/hooks/useEmailVerified';
@@ -15,6 +14,7 @@ import useShowTrialWelcomeModal from 'src/hooks/useShowTrialWelcomeModal';
 import WelcomeModal from 'src/components/welcome/WelcomeModal';
 import { FeatureGate } from 'src/components/common/FeatureGate';
 import { Product } from 'boclips-api-client/dist/sub-clients/accounts/model/Account';
+import FeaturedPlaylists from 'src/components/carousel/FeaturedPlaylists';
 import FilmIcon from '../../resources/icons/film-icon.svg';
 import YourLibraryIcon from '../../resources/icons/your-library.svg';
 import s from './style.module.less';
